+++ include/game_utils.hpp
@@ -31,7 +31,7 @@ void UpdateEnemyAnimation(Enemy* enemy);
 void UpdateProjectileAnimation(Projectile* projectile);
 float CalculateMagnitude(SDL_Rect a, SDL_Rect b);
 void CalculateDirection(SDL_Rect a, SDL_Rect b, Projectile* projectile);
-void FireSingleProjectile(SDL_Rect player_rect, SDL_Texture* texture, int speed, int total_frames,
+void FireSingleProjectile(SDL_Rect player_rect, SDL_Texture* texture, float speed, int total_frames,
     SDL_Rect rect_src, SDL_Rect rect_dst,
     int delay, Uint32& last_projectile_time, Mix_Chunk* sound, projectileType type, int animation_speed);
 Enemy* FindClosestEnemy(SDL_Rect player_rect, Enemy enemies[], int max_enemies);
